+++ src/view/board.js
@@ -1,0 +1,10 @@
+const createBoardMarkup = () => {
+  return(
+    `<section class="board container">
+      <div class="board__tasks"></div>
+    </section>`
+  );
+}
+
+
+export { createBoardMarkup }
