+++ src/view/more-button.js
@@ -1,0 +1,5 @@
+const createMoreButtonMarkup = () => {
+  return(`<button class="load-more" type="button">load more</button>`);
+}
+
+export { createMoreButtonMarkup }
