+++ src/view/task.js
@@ -1,0 +1,41 @@
+const createTaskMarkup = () => {
+  return(
+    `<article class="card card--yellow">
+          <div class="card__form">
+            <div class="card__inner">
+              <div class="card__control">
+                <button type="button" class="card__btn card__btn--edit">
+                  edit
+                </button>
+                <button type="button" class="card__btn card__btn--archive">
+                  archive
+                </button>
+                <button
+                  type="button"
+                  class="card__btn card__btn--favorites card__btn--disabled"
+                >
+                  favorites
+                </button>
+              </div>
+
+              <div class="card__color-bar">
+                <svg class="card__color-bar-wave" width="100%" height="10">
+                  <use xlink:href="#wave"></use>
+                </svg>
+              </div>
+
+              <div class="card__textarea-wrap">
+                <p class="card__text">Example task with custom color and without date.</p>
+              </div>
+
+              <div class="card__settings">
+                <div class="card__details">
+                </div>
+              </div>
+            </div>
+          </div>
+        </article>`
+  );
+}
+
+export { createTaskMarkup }
